+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `--type` option to `ig_markets search` to only show instruments of specific type(s)
 - Use a default currency of `USD` for `ig_markets orders create` and `ig_markets positions create`
 - Improved command-line client help output for several commands
+- Added `IGMarkets::DealingPlatform::SprintMarketPositionMethods#[]`
 - Fixed `IGMarkets::DealingPlatform::PositionMethods#[]` throwing on an unknown deal ID
 
 ### 0.5 — April 23, 2016